update-maria-db.md for updating mariadb server
Script that updates MariaDB to the desired version. Keep in mind that updating is only possible to a higher version than the currently installed one. This is an unofficial script, that can cause unexpected behaviors, so run it at your own risk. After running, it may say error, but it is most likely only a warning because a source for MariaDB already existed. In case anything really fails, a backup of your DB's are at /var/lib/mysql_backup